auto database updation and deletion added
