+++ tweepy-bot/database.py
@@ -22,21 +22,24 @@ except Exception as e:
     
 c = conn.cursor()
 
+def data_count():
+    '''
+    Returns the number of rows in the database
+    '''
+    x = c.execute("SELECT * FROM issue")
+    conn.commit()
+    count = len(c.fetchall())
+    return count
 
-def convertTuple(tup): 
+
+def convertTuple(tup):
+    '''
+    convert a tuple to string
+    '''
     str =  ''.join(tup) 
     return str
 
-def insert_data(issue_title,issue_url):
-    '''
-    takes the issue title and issue url and inserts into the database
-    '''
-    try:
-        c.execute("INSERT INTO issue (issue_text,url) VALUES (?,?)", (issue_title,issue_url))
-        conn.commit()
-    except sqlite3.IntegrityError:
-        pass
-        logger.info("Database Updated!")
+
 
 
 def create_schema():
@@ -46,12 +49,24 @@ def create_schema():
     '''
     
     c.execute("""CREATE TABLE IF NOT EXISTS issue (
-            ID INTEGER PRIMARY KEY AUTOINCREMENT,
+            ID INTEGER PRIMARY KEY,
             issue_text TEXT,
             url TEXT NOT NULL UNIQUE
             )""")
     conn.commit()
     logger.info("Database Created!")
+
+def insert_data(issue_title,issue_url):
+    '''
+    takes the issue title and issue url and inserts into the database
+    '''
+    try:
+        c.execute("INSERT INTO issue (issue_text,url) VALUES (?,?)", (issue_title,issue_url))
+        conn.commit()
+        logger.info("Database Updated, New issue inserted!")
+    except sqlite3.IntegrityError:
+        pass
+        logger.info("Dublicate Data, not inserted")
 
 def get_last_data():
     '''
@@ -67,14 +82,13 @@ def del_last_data():
     '''
     Deletes the latest row in the database
     '''
-    x = c.execute("SELECT * FROM issue")
-    count = len(c.fetchall())
+    count = data_count()
     try:
-        c.execute("DELETE FROM issue WHERE rowid = (?)", (count,))
+        c.execute("DELETE FROM issue WHERE ROWID = :ct", {"ct":count})
+        conn.commit()
         print("Data deleted")
     except:
         pass
         print("No more data in database")
-    conn.commit()
